Refactor Loading
