+++ frontend/src/components/molecules/Loading.tsx
@@ -1,14 +1,11 @@
 import React from "react";
+import CircularProgress from "@mui/material/CircularProgress";
 
 function Loading() {
   return (
     <div className="h-screen bg-white">
       <div className="flex justify-center items-center h-full">
-        <img
-          className="h-8 w-8"
-          src="https://icons8.com/preloaders/preloaders/1488/Iphone-spinner-2.gif"
-          alt=""
-        />
+        <CircularProgress />
       </div>
     </div>
   );
